+++ CHANGELOG.md
@@ -6,8 +6,13 @@
 - Added base setup for Animation Nodes' extension.
 - Added build and stale workflows.
 - Added *Material Instancer* node.
+- Added *Texture Input* node.
+- Added *Texture Falloff* node.
+- Added Extra Nodes category in node menu.
 
 ### Fixed
 
 
 ### Changed
+
+- Update the `__init__.py` code.